Move static methods out of completer to helper
