+++ lib/pry/helpers.rb
@@ -1,5 +1,6 @@
 require "pry/helpers/base_helpers"
 require "pry/helpers/options_helpers"
 require "pry/helpers/command_helpers"
+require "pry/helpers/completer_helpers"
 require "pry/helpers/text"
 require "pry/helpers/table"
